refactor: moved Person object initiation into separate file
